requirements.txt
pyzipper~=0.3.6
tqdm~=4.67.0
pillow~=11.0.0